Moved project structure to gradle
The project is now jitpack compatible

Added VersionChecker
