+++ src/main/java/de/dosmike/sponge/mikestoolbox/database/H2LocationSerializer.java
