+++ src/main/java/de/dosmike/sponge/mikestoolbox/event/BoxSneakEvent.java
