+++ src/main/java/de/dosmike/sponge/mikestoolbox/living/CustomEffect.java
