+++ src/main/java/de/dosmike/sponge/mikestoolbox/nbtinspect/BoxNbtInspector.java
